Use Feature::Compat::Try to catch errors
This is an improvement over the previous change for GH#2, although there
is a possible incompatability by including a file path and line number
in the error message.

Note that we use Feature::Compat::Try instead of Try::Tiny, which is
already used by Catalyst.  The main reason is that it is forward-
compatible with newer Perls.