Re-add the file existence check that was removed in v0.1.3.
This effectively reverts a33b32bdba4be4add13430561e41913558e4abfd.

Technically the check for file existence adds a race condition when the
file is deleted between the existence check and the file opening.  It
should be redunant, since the file opening will fail with an error
message that the file was not found.

However, there are CPAN Tester failures where the process seems to exit
quickly, and the error is never caught.

It's also worth noting that there is no problem with the extra check
before opening the file.  If the file is somehow deleted before opening,
then the failure will be caught anyway.